refactored into smaller components
